Controlled gates now work. Added more gates. Starting documentation.

+++ src/math/matrix.hpp
@@ -21,27 +21,27 @@ class matrix : public xobject {
 
     private:
         std::vector<complex> values;
-        uint rows;
-        uint columns;
-        uint length;
+        size_t rows;
+        size_t columns;
+        size_t length;
 
     public:
         /// </Summary>
         /// Create an empty matrix of this size
         /// </Summary>
-        matrix(int rows, int columns){
+        matrix(size_t rows, size_t columns){
             this->rows = rows > 0 ? rows : 1;
             this->columns = columns > 0 ? columns : 1;
             this->length = (this->rows) * (this->columns);
             this->values = std::vector<complex>(this->length);
-            for(uint i = 0; i < this->length; i++)
+            for(size_t i = 0; i < this->length; i++)
                 this->values[i] = complex();
         }
 
         /// </Summary>
         /// Create a matrix of this size with the given values in compressed row->column form 
         /// </Summary>
-        matrix(int rows, int columns, std::vector<complex> values) : matrix(rows, columns){
+        matrix(size_t rows, size_t columns, std::vector<complex> values) : matrix(rows, columns){
             size_t smallest = this->length < values.size() ? this->length : values.size();
             for(size_t i = 0; i < smallest; i++){
                 this->values[i] = values[i];
@@ -58,8 +58,8 @@ class matrix : public xobject {
         /// </Summary>
         /// Get element from matrix
         /// </Summary>
-        complex& operator() (int r, int c) {
-            int idx = r * columns + c;
+        complex& operator() (size_t r, size_t c) {
+            size_t idx = r * columns + c;
             if(idx < 0 || idx >= length){
                 throw std::out_of_range("Row or column is out of range");
             }
@@ -69,7 +69,7 @@ class matrix : public xobject {
         /// </Summary>
         /// Get element from matrix
         /// </Summary>
-        complex& operator() (int idx) {
+        complex& operator() (size_t idx) {
             if(idx < 0 || idx >= length){
                 throw std::out_of_range("Row or column is out of range");
             }
@@ -79,29 +79,29 @@ class matrix : public xobject {
         /// </Summary>
         /// Number of rows
         /// </Summary>
-        uint countRows(){
+        size_t countRows(){
             return this->rows;
         }
 
         /// </Summary>
         /// Number of columns
         /// </Summary>
-        uint countColumns(){
+        size_t countColumns(){
             return this->columns; 
         }
 
         /// </Summary>
         /// Total number of elements
         /// </Summary>
-        uint size(){
+        size_t size(){
             return this->length;
         }
 
         /// </Summary>
         /// Set element in matrix
         /// </Summary>
-        void set(int r, int c, complex value){
-            int idx = r * columns + c;
+        void set(size_t r, size_t c, complex value){
+            size_t idx = r * columns + c;
             if(idx < 0 || idx >= length){
                 throw std::out_of_range("Row or column is out of range");
             }
@@ -111,7 +111,7 @@ class matrix : public xobject {
         /// </Summary>
         /// Set element in matrix
         /// </Summary>
-        void set(int idx, complex value){
+        void set(size_t idx, complex value){
             if(idx < 0 || idx >= length){
                 throw std::out_of_range("Row or column is out of range");
             }
@@ -121,8 +121,8 @@ class matrix : public xobject {
         /// </Summary>
         /// Get element from matrix
         /// </Summary>
-        complex get(int r, int c){
-            int idx = r * columns + c;
+        complex get(size_t r, size_t c){
+            size_t idx = r * columns + c;
             if(idx < 0 || idx >= length){
                 throw std::out_of_range("Row or column is out of range");
             }
@@ -132,7 +132,7 @@ class matrix : public xobject {
         /// </Summary>
         /// Get element from matrix
         /// </Summary>
-        complex get(int idx){
+        complex get(size_t idx){
             if(idx < 0 || idx >= length){
                 throw std::out_of_range("Row or column is out of range");
             }
@@ -145,11 +145,11 @@ class matrix : public xobject {
         std::string toString(){
             std::stringstream sb;
             sb << "{";
-            for(uint r = 0; r < this->rows; r++){
+            for(size_t r = 0; r < this->rows; r++){
                 if(r != 0)
                     sb << ";";
                 sb << "{";
-                for(uint c = 0; c < this->columns; c++){
+                for(size_t c = 0; c < this->columns; c++){
                     if(c != 0)
                         sb << ",";
                     sb << get(r, c).toString();
@@ -171,7 +171,7 @@ matrix operator + (matrix a, matrix b){
         throw std::length_error("Matrix dims do not match");
     }
     matrix m(a.countRows(), a.countColumns());
-    for(uint i = 0; i < a.size(); i++){
+    for(size_t i = 0; i < a.size(); i++){
         m.set(i, a.get(i) + b.get(i));
     }
     return m;
@@ -182,7 +182,7 @@ matrix operator - (matrix a, matrix b){
         throw std::length_error("Matrix dims do not match");
     }
     matrix m(a.countRows(), a.countColumns());
-    for(uint i = 0; i < a.size(); i++){
+    for(size_t i = 0; i < a.size(); i++){
         m.set(i, a.get(i) - b.get(i));
     }
     return m;
@@ -190,7 +190,7 @@ matrix operator - (matrix a, matrix b){
 
 matrix operator * (complex s, matrix a){
     matrix m(a.countRows(), a.countColumns());
-    for(uint i = 0; i < a.size(); i++){
+    for(size_t i = 0; i < a.size(); i++){
         m.set(i, a.get(i) * s);
     }
     return m;
@@ -198,7 +198,7 @@ matrix operator * (complex s, matrix a){
 
 matrix operator * (matrix a, complex s){
     matrix m(a.countRows(), a.countColumns());
-    for(uint i = 0; i < a.size(); i++){
+    for(size_t i = 0; i < a.size(); i++){
         m.set(i, a.get(i) * s);
     }
     return m;
@@ -206,7 +206,7 @@ matrix operator * (matrix a, complex s){
 
 matrix operator / (matrix a, complex s){
     matrix m(a.countRows(), a.countColumns());
-    for(uint i = 0; i < a.size(); i++){
+    for(size_t i = 0; i < a.size(); i++){
         m.set(i, a.get(i) / s);
     }
     return m;
@@ -218,10 +218,10 @@ matrix operator * (matrix a, matrix b){
     }
     matrix m(a.countRows(), b.countColumns());
 
-    for(uint i = 0; i < a.countRows(); i++){
-        for(uint j = 0; j < b.countColumns(); j++){
+    for(size_t i = 0; i < a.countRows(); i++){
+        for(size_t j = 0; j < b.countColumns(); j++){
             complex sum = 0;
-            for(uint k = 0; k < a.countColumns(); k++){
+            for(size_t k = 0; k < a.countColumns(); k++){
                 sum = sum + a(i,k) * b(k,j);
             }
             m.set(i,j, sum);
@@ -234,19 +234,19 @@ matrix operator * (matrix a, matrix b){
 //Tensor product
 matrix operator << (matrix a, matrix b){
 	//Compute width
-	uint nw = a.countColumns() * b.countColumns();
-	uint nh = a.countRows() * b.countRows();
+	size_t nw = a.countColumns() * b.countColumns();
+	size_t nh = a.countRows() * b.countRows();
 		
 	//Create matrix;
 	matrix m(nh, nw);
 	
 	//Loop over all elements setting as necessary
-	uint r = 0;
-    uint c = 0;
-    for(uint ar = 0; ar < a.countRows(); ar++){
-        for(uint ac = 0; ac < a.countColumns(); ac++){
-            for(uint br = 0; br < b.countRows(); br++){
-                for(uint bc = 0; bc < b.countColumns(); bc++){
+	size_t r = 0;
+    size_t c = 0;
+    for(size_t ar = 0; ar < a.countRows(); ar++){
+        for(size_t ac = 0; ac < a.countColumns(); ac++){
+            for(size_t br = 0; br < b.countRows(); br++){
+                for(size_t bc = 0; bc < b.countColumns(); bc++){
                     m(br + ar * b.countRows(), bc + ac * b.countColumns()) = a(ar, ac) * b(br, bc); 
                 }
             }
